task: semantic table component
